+++ packages/analytics/dashboard-renderer/src/components/DonutChartRenderer.vue
@@ -11,8 +11,8 @@
 
 <script setup lang="ts">
 import BaseAnalyticsChartRenderer from './BaseAnalyticsChartRenderer.vue'
-import type { RendererProps } from '../types'
+import type { ChartRendererProps } from '../types'
 import type { DonutChartOptions } from '@kong-ui-public/analytics-utilities'
 
-defineProps<RendererProps<DonutChartOptions>>()
+defineProps<ChartRendererProps<DonutChartOptions>>()
 </script>