+++ packages/core/table-data-grid/src/components/TableDataGrid.vue
@@ -151,11 +151,13 @@ const onGridReady = (event: GridReadyEvent<Row>) => {
 
 <style lang="scss" scoped>
 .kong-ui-public-table-data-grid {
-  border: 1px solid var(--kui-color-border, $kui-color-border);
+  border: none;
   border-radius: var(--kui-border-radius-20, $kui-border-radius-20);
+  box-sizing: border-box;
   display: flex;
   flex-direction: column;
-  min-height: 360px;
+  height: 100%;
+  min-height: 0;
   overflow: hidden;
   width: 100%;
 }
@@ -166,13 +168,15 @@ const onGridReady = (event: GridReadyEvent<Row>) => {
   --ag-border-color: var(--kui-color-border, #{$kui-color-border});
   --ag-header-background-color: var(--kui-color-background, #{$kui-color-background});
   --ag-header-column-border: 1px solid var(--kui-color-border, #{$kui-color-border});
+  --ag-header-column-border-height: 30%;
   --ag-header-column-resize-handle-color: transparent;
+  --ag-header-font-weight: var(--kui-font-weight-semibold, #{$kui-font-weight-semibold});
   --ag-wrapper-border: none;
   --ag-wrapper-border-radius: 0;
   /* stylelint-enable custom-property-pattern */
 
-  flex: 1 1 360px;
-  min-height: 360px;
+  flex: 1 1 auto;
+  min-height: 0;
   width: 100%;
 }