+++ packages/entities/entities-plugins/src/components/PluginEntityForm.vue
@@ -25,6 +25,8 @@
         :plugin-name="formModel.name"
         :render-rules="pluginConfig?.renderRules"
         :schema="freeformSchema"
+        @click:create-entity="(payload: EntityCreateEvent) => $emit('click:create-entity', payload)"
+        @click:learn-more="(entity: string) => $emit('click:learn-more', entity)"
         @global-action="(name: GlobalAction, payload: any) => $emit('globalAction', name, payload)"
       >
         <template
@@ -160,6 +162,8 @@ const emit = defineEmits<{
   ): void
   (e: 'showNewPartialModal', redisType: string): void
   (e: 'globalAction', name: GlobalAction, payload: any): void
+  (e: 'click:learn-more', entity: string): void
+  (e: 'click:create-entity', payload: EntityCreateEvent): void
   (e: 'validity-change', payload: PluginValidityChangeEvent): void
   (e: 'click:create-entity', payload: EntityCreateEvent): void
   (e: 'click:learn-more', entity: string): void