+++ packages/entities/entities-plugins/src/components/fields/kong-identity/ConfigFormWithIdentity.vue
@@ -1,6 +1,9 @@
 <template>
   <DynamicLayout v-bind="props">
-    <ConfigFormContent />
+    <ConfigFormContent
+      @click:create-entity="(payload) => emit('click:create-entity', payload)"
+      @click:learn-more="(entity: string) => emit('click:learn-more', entity)"
+    />
   </DynamicLayout>
 </template>
 
@@ -11,8 +14,14 @@ import DynamicLayout from '../../free-form/shared/layout/DynamicLayout.vue'
 import ConfigFormContent from './ConfigFormContent.vue'
 
 import type { PluginFormLayoutProps as Props } from '../../free-form/shared/layout/provider'
+import type { EntityCreateEvent } from '../../../types'
 
 const props = defineProps<Props>()
+
+const emit = defineEmits<{
+  'click:learn-more': [entity: string]
+  'click:create-entity': [payload: EntityCreateEvent]
+}>()
 
 const slots = defineSlots<{
   [K in typeof AUTOFILL_SLOT_NAME]: () => any