+++ packages/entities/entities-redis-configurations/src/components/RedisConfigurationConfigCard.vue
@@ -116,12 +116,12 @@
                   <ChevronUpIcon
                     v-if="!isCollapsed"
                     decorative
-                    :size="KUI_ICON_SIZE_40"
+                    :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`"
                   />
                   <ChevronRightIcon
                     v-else
                     decorative
-                    :size="KUI_ICON_SIZE_40"
+                    :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`"
                   />
                   {{ partialCollapseTriggerLabel }}
                 </KButton>