+++ packages/core/core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/core@1.11.26...@kong-ui-public/core@1.11.27) (2026-06-22)
+
+
+### Bug Fixes
+
+* design-tokens eslint and stylelint ([#3477](https://github.com/Kong/public-ui-components/issues/3477)) ([700ad19](https://github.com/Kong/public-ui-components/commit/700ad194ac6a2b9c753c018c4db9319f1cd50232))
+
+
+
+
+
 ## [1.11.26](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/core@1.11.25...@kong-ui-public/core@1.11.26) (2026-06-11)